Make event route return an event model in model hook.
Previously, the model hook returned an attendee model object with an association to the event. The component mutated the attendee model object and eventually it was saved into the store. 

It makes more sense to make event the model, given the route, pathname and ui of this page. The attendee is now built as a POJO object in the new-attendee-form component. The attendee model object is not created until it’s about to be saved into the store, initiated by an action sent via a user interaction.