Handle errors for form completion.
Create a new form-message component to encapsulate the messages in the process. Decided after a few tries to move the logic to a bunch of “if’s” in the form-message template because it is the clearest. Though it is a little repetitive, I think it’s clearer than relying on a bunch of switch statements in computed properties to tell what the text, class, and img is.